+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [10686](https://github.com/vegaprotocol/vega/issues/10764) - Network will now only trade within price monitoring bounds.
 - [10727](https://github.com/vegaprotocol/vega/issues/10727) - Allow for a 0 funding rate scaling factor.
+- [10770](https://github.com/vegaprotocol/vega/issues/10770) - Increase price monitoring triggers to 100.
 
 ## 0.74.6